orchestration/requirements.txt
# Apache Airflow with constraints
apache-airflow==2.8.0
apache-airflow[postgres,redis,celery,crypto,ssh]==2.8.0

# Airflow AWS providers
apache-airflow-providers-amazon==8.12.0

# DBT Core with adapters
dbt-core==1.7.4
dbt-duckdb==1.7.2
dbt-athena-community==1.7.2

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Database drivers
psycopg2-binary==2.9.9
redis==5.0.1

# Utilities
pyyaml==6.0.1
requests==2.31.0
pandas==2.1.4
sqlalchemy==1.4.46

# Development and testing
pytest==7.4.4
pytest-airflow==0.0.8

# Logging and monitoring
structlog==23.2.0

# Security
cryptography==41.0.8